Added a GUI
Still need to make it efficient. The problem right now is that the user can't directly control the number of pages requests. Need to add an option for that. Then make sure to make the searching on a different thread so that way it doesn't affect the GUI. Need to add an "open in web browser" button for a selected item at the end